tabula/requirements.txt
asgiref==3.8.1
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
JPype1==1.5.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tabula-py==2.9.0
tzdata==2024.1
